Move template-rendering helpers to separate module
